Collect ItemService tests in a single file
Delete trivial create-config test
