+++ packages/sdk/test/fixtures/outputs/get-item.output.json
@@ -1,23 +1,23 @@
 {
-  "id": null,
-  "own": null,
+  "id": "itemId",
+  "own": true,
   "name": "food",
   "label": "My Fave Foods",
-  "description": null,
+  "description": "",
   "created_at": "1970-01-01T00:00:00.001Z",
-  "item_template_id": null,
-  "ordinal": null,
-  "visible": null,
+  "item_template_id": "id",
+  "ordinal": 1,
+  "visible": true,
   "updated_at": "1970-01-01T00:00:00.001Z",
-  "item_template_label": null,
+  "item_template_label": "",
   "image": null,
   "item_image": null,
   "item_image_background_colour": null,
-  "classification_node_ids": null,
-  "association_ids": null,
-  "associations_to_ids": null,
+  "classification_node_ids": [],
+  "association_ids": [],
+  "associations_to_ids": [],
   "slot_ids": ["steak", "pizza", "yoghurt"],
-  "me": null,
+  "me": true,
   "background_color": null,
   "original_id": null,
   "owner_id": null,
@@ -25,17 +25,17 @@
   "slots": [
     {
       "id": "pizza",
-      "own": null,
+      "own": true,
       "share_id": null,
       "name": "pizza",
       "description": null,
       "encrypted": false,
-      "ordinal": null,
-      "visible": null,
-      "classification_node_ids": null,
+      "ordinal": 1,
+      "visible": true,
+      "classification_node_ids": [],
       "attachment_id": null,
-      "item_id": null,
-      "required": null,
+      "item_id": "itemId",
+      "required": false,
       "updated_at": "1970-01-01T00:00:00.001Z",
       "created_at": "1970-01-01T00:00:00.001Z",
       "slot_type_name": "key_value",
@@ -52,17 +52,17 @@
     },
     {
       "id": "steak",
-      "own": null,
+      "own": true,
       "share_id": null,
       "name": "steak",
       "description": null,
       "encrypted": false,
-      "ordinal": null,
-      "visible": null,
-      "classification_node_ids": null,
+      "ordinal": 1,
+      "visible": true,
+      "classification_node_ids": [],
       "attachment_id": null,
-      "item_id": null,
-      "required": null,
+      "item_id": "itemId",
+      "required": false,
       "updated_at": "1970-01-01T00:00:00.001Z",
       "created_at": "1970-01-01T00:00:00.001Z",
       "slot_type_name": "key_value",
@@ -79,17 +79,17 @@
     },
     {
       "id": "beer",
-      "own": null,
+      "own": true,
       "share_id": null,
       "name": "beer",
       "description": null,
       "encrypted": false,
-      "ordinal": null,
-      "visible": null,
-      "classification_node_ids": null,
+      "ordinal": 1,
+      "visible": true,
+      "classification_node_ids": [],
       "attachment_id": null,
-      "item_id": null,
-      "required": null,
+      "item_id": "itemId",
+      "required": false,
       "updated_at": "1970-01-01T00:00:00.001Z",
       "created_at": "1970-01-01T00:00:00.001Z",
       "slot_type_name": "key_value",
